Rename Parking Full to Parking Application
Parking Application: Allotting Parking spaces of the mall with specific parking numbers. With the help of sensors and interfacing it with the application parking status of the slot can be checked on the application. online booking of the parking space for certain period can be done by online payment while booking a movie ticket also. A person can book a parking slot and can pay while taking the token after reaching mall. This feature will be complemented with the led lights on the floor which will guide the car to the parking spot booked by it. Benefits of this feature are: 1) Systematic Parking with optimized use of the space. 2) Time and Efforts are reduced for parking. 3)Proper record will be kept of parked cars. 4)Easy to find the booked place as the LED lights will guide to the place. 5)Sometimes customer doesn't gets a parking place due to parking full which leads to sometimes waste their time or unable to watch movie at booked time. 6)Difficult to pay change so online payment will ease the payment process.